Big routing changes. Can select deck from list on home screen and then cycle through the cards in that deck in the review screen. Back button now sends back to home screen
