+++ prisma/schema.prisma
@@ -13,22 +13,23 @@ generator client {
 }
 
 model DxDeck {
-  id              Int       @default(autoincrement()) @id
+  id              String    @id @default(uuid())
   createdAt       DateTime  @default(now())
   updatedAt       DateTime  @updatedAt
   type            String    @db.VarChar(255)
+  name            String    @db.Text
   cards           DxCard[]
 }
 
 model DxCard {
-  id              Int       @default(autoincrement()) @id
+  id              String    @id @default(uuid())
   createdAt       DateTime  @default(now())
   updatedAt       DateTime  @updatedAt
   front           String    @db.Text
   back            String    @db.Text
   source          String    @db.VarChar(255)
   xp              Float     @db.Float 
-  cardId          Int
-  deck            DxDeck    @relation(fields: [cardId], references: [id], onDelete: Cascade)
-  @@index([cardId])
+  deckId          String
+  deck            DxDeck    @relation(fields: [deckId], references: [id], onDelete: Cascade)
+  @@index([deckId])
 }
